MELOSYS-8084 Rett funn fra code review
Flytt query-fabrikken til httpClients (repo-konvensjon, gjenbruker
API_PROXY_URL), retry: false som øvrige queries, guard mot tom
toggle-liste, og fjern ubrukt barrel-fil.
