--- app/src/featuretoggle/index.ts
@@ -1,2 +1,0 @@
-export * from "./toggleNavn.ts";
-export * from "./useFeatureToggle.ts";
